% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchExam.R
\name{batchExam}
\alias{batchExam}
\title{Examine batch effects}
\usage{
batchExam(
  prepsce,
  xdim = 10,
  ydim = 10,
  k = 20,
  seed = 1234,
  downsample = 2000,
  DR_method = "UMAP"
)
}
\arguments{
\item{prepsce}{a SingleCellExperiment object generated by batchPrep function}

\item{xdim}{integer, parameter for flowSOM clustering}

\item{ydim}{integer, parameter for flowSOM clustering}

\item{k}{integer, parameter for flowSOM clustering}

\item{seed}{numeric vector; the seed set for each run of clustering}

\item{downsample}{integer specifying the number of cells per sample for clustering}

\item{DR_method}{character string specifying the method for dimensionality reduction}
}
\value{
a ggplot object
}
\description{
Examine variation across batches on clustering and cluster proportion level, please refer to Mtrussart et al(CytofRUV)
for details.
}
\examples{
#select TSNE for DR with default settings
batchExam(prepsce,method='TSNE')
#select UMAP for DR with 3000 cells per sample
batchExam(prepsce,k=30,downsample=3000,DR_method='TSNE')

}
