% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchNorm.R
\name{batchNorm}
\alias{batchNorm}
\title{Batch correction to remove unwanted non-biological variation among dataset}
\usage{
batchNorm(
  fcs_dir = NULL,
  md_dir,
  panel_dir,
  method = "CytoNorm",
  nCells = 6000,
  xdim = 5,
  ydim = 5,
  nClus = 10,
  nQ = 101,
  goal = "mean",
  k = 5,
  output_dir = "Norm_output"
)
}
\arguments{
\item{fcs_dir}{file path of the fcs files}

\item{md_dir}{file path of the metadata file}

\item{panel_dir}{file path of the panel data file}

\item{method}{character string specifying the correction method to be used, default is 'CytoNorm', other option is 'CytofRUV'}

\item{nCells}{integer, CytoNorm parameter to specify the number of cells for training and model,default is 6000}

\item{xdim}{integer, CytoNorm parameter to specify the xdim of flowSOM clustering, default is 5}

\item{ydim}{integer,CytoNorm parameter to specify the ydim of flowSOM clustering, default is 5}

\item{nClus}{integer, CytoNorm and CytofRUV parameter to specify the k number of flowSOM clustering, default is 10}

\item{nQ}{integer, CytoNorm parameter to specify the number of quantiles, default is 101}

\item{goal}{character string, CytoNorm parameter, default is 'mean'}

\item{k}{integer,CytofRUV parameters,default is 5}

\item{output_dir}{file path of the output files(after normalisation),default is 'Norm_output'folder under working directory}
}
\value{
fcs files after normalisation
}
\description{
Batch normalization using CytoNorm or CytofRUV to remove unwanted variation arising from
experimental or instrumental artifacts. Please refer to Mtrussart et al.(CytofRUV) and Gassen et al.(CytoNorm)
to set relevant parameters.
}
\examples{
#set the path for fcs files, sample metadata and panel metadata
fcs_dir <- 'fcs'
metadata_dir <- 'fcs/metadata2.xlsx'
panel_dir <- 'fcs/panel_metadata2.xlsx'
#Use the default CytoNorm method and parameters for batch normalisation
batchNorm(fcs_dir,metadata_dir,panel_dir)
#Still use CytoNorm method, but change k number(nClus) as 20 for flowSOM clustering
batchNorm(fcs_dir,metadata_dir,panel_dir,nClus=20)
#Use CytofRUV method, set nClus as 20 and k value as 6
batchNorm(fcs_dir,metadata_dir,panel_dir,method='CytofRUV',nClus=20,k=6)

}
