% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchPrep.R
\name{batchPrep}
\alias{batchPrep}
\title{Prepare data for batch effect examination}
\usage{
batchPrep(fcs_dir, md_dir, panel_dir)
}
\arguments{
\item{fcs_dir}{file path for fcs files}

\item{md_dir}{file path for sample metadata}

\item{panel_dir}{file path for panel metadata}
}
\value{
a sce object
}
\description{
By loading sample metadata, panel metadata and fcs files to construct a sce object for batch effect examination and normalisation
}
\examples{
fcs_dir <- 'fcs'
metadata_dir <- 'fcs/metadata3.xlsx'
panel_dir <- 'fcs/panel_metadata2.xlsx'
prepsce <- batchPrep(fcs_dir,metadata_dir,panel_dir)
}
