% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryClass.R
\name{binaryClass}
\alias{binaryClass}
\title{Performing binary classification}
\usage{
binaryClass(data = NULL, class.file = NULL, scale = FALSE)
}
\arguments{
\item{data}{data frame of marker expression, which can be exported from sce object}

\item{class.file}{file path to user pre-defined classification matrix(a csv file)}

\item{scale}{logical vector, standarisation of data, default is F}
}
\value{
table of the classification/defination results of each cell
}
\description{
Binary classification and annotation will be performed based on user-defined classification matrix
}
\examples{
binary_results <- binaryClass(exprs_n, class_dir)
}
