% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayMarkers.R
\name{displayMarkers}
\alias{displayMarkers}
\title{Display marker intensity distribution selected for Binary clustering}
\usage{
displayMarkers(sce, type, color_by = "sample_id", ncol = 3)
}
\arguments{
\item{sce}{a SingleCellExpriment object}

\item{type}{a data frame of user pre-defined classification matrix}

\item{color_by}{character string specifying the color coding, default is 'sample_id'}

\item{ncol}{a numeric vector specifying the number of subplot per row, default is 3}
}
\value{
a ggplot object
}
\description{
Display marker intensity distribution selected for Binary clustering
}
\examples{
#input/read types file as a data frame
class_dir <- 'types.csv'
types <- read_csv('types.csv')

#construct a sce object
sce <- prepData(fs, panel, md, features = panel$fcs_colname)

#show marker intensity distribution of the type file(classification matrix) with 4 subplots per row, colored by study conditions
displayMarkers(sce,types,color_by='condition',n=4)

}
