% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCluster.R
\name{extractCluster}
\alias{extractCluster}
\title{Extract a certain cell cluster of interest from the total population}
\usage{
extractCluster(sce, cluster, null_cluster_id = T)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{cluster}{character string specifying the cell cluster to extract}

\item{null_cluster_id}{logical vector to determine whether to clear out the cluster_id in the cluster to be extracted,
default is 'T', which is also recommended, because further interrogation/clustering might be done in the new cluster}
}
\value{
a SingleCellExperiment object for the extracted cluster
}
\description{
Extract a certain cell cluster of interest from the total population
}
\examples{
# to extract B cells from total cells, and define it to a new sce object named b_sce
b_sce<-extractCluster(sce,cluster='B Cells',null_cluster_id = F)
}
