% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqData.R
\name{freqData}
\alias{freqData}
\title{Extract raw data(table) of cell frequencies from a sce object}
\usage{
freqData(sce, metadata = T)
}
\arguments{
\item{sce}{a SingleCellExperiment objcet}

\item{metadata}{logical vector, default is T which means meatadata will be added into the table}
}
\value{
a table with sample_id, cluster_id, frequency, and metadata(e.g.condition, outcome) if 'metadata=T'.
}
\description{
Directly extract raw data of cell frequencies from a sce object in order to perform more customized analysis dependent on
the research questions
}
\examples{

fq<-freqData(sce,metadata=T)
}
