% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotbcFreq.R
\name{plotbcFreq}
\alias{plotbcFreq}
\title{Plot total cell population frequencies after binary classification}
\usage{
plotbcFreq(sce, binary.results, style = "bar")
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{binary.results}{vector generated by running binaryClass}

\item{style}{character string indicating the type of the plot, options are 'bar' or 'pie', default is 'bar'}

\item{binary_results}{list, results of binary classification}

\item{data}{sce object}
}
\value{
a ggplot object

a data frame
}
\description{
Visualisation of cell population frequencies/composition after binary classification, in the format of bar plot
}
\examples{
plotbcFreq(sce,binary.results,style='pie')

Title

}
