% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFreq.R
\name{plotFreq}
\alias{plotFreq}
\title{Plot of cell population frequencies across samples, conditions or clusters}
\usage{
plotFreq(sce, type = "box", title = "Cell abundances", group_by)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{type}{character string specifying the type of plot, options are 'stacked' indicating
stacked histogram, and 'box' indicating boxplot;default is 'box'}

\item{title}{character string, title of the plot; default is 'Cell Abundance'}

\item{group_by}{character string specifying the metadata used to group by; valid values are for instance 'condition'etc.}
}
\value{
a ggplot object
}
\description{
Visualization of cell population frequencies across samples or conditions after clustering
in the format of boxplot or stacked histogram
}
\examples{
#plot a boxplot grouped by condition
plotFreq(sce,title='Cell Abundance',group_by='condition')
#plot a stacked histogram grouped by batch
plotFreq(sce,type='stacked',title='Cell Abundance',group_by='batch')

}
