% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStateMarkerExprs.R
\name{plotStateMarkerExprs}
\alias{plotStateMarkerExprs}
\title{Plot expression intensity of the state markers}
\usage{
plotStateMarkerExprs(sce, group)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{group}{character string specifying the parameter for comparison}
}
\value{
a ggplot object and table for statistical analysis
}
\description{
Plot state marker expression intensity of each cluster
}
\examples{
plotStateMarkerExprs(sce,group='condition')
}
