% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPheno.R
\name{runPheno}
\alias{runPheno}
\title{Run Phenograph algorithm for clustering}
\usage{
runPheno(sce, k, downsample = NULL, seed = 1234, type_markers = T)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{k}{integer; number of nearest neighbors}

\item{downsample}{logical/numeric vector; default is F which means the input data will not be downsampled;
Or input can be a number to specify how many cells per sample to be clustered;}

\item{seed}{numeric vector; the seed set for each run of clustering}

\item{type_markers}{logical vector; default is NULL, which means only type markers will be used for clustering;if it is F,
then all markers will be used for clustering}
}
\value{
a SingleCellExpreriment with newly added data of Phenograph clustering results as below:

-colData: cluster_id of each cell inferred by Phenograph

-metadata:cluster_codes/rph, stores the clustering levels generated by Phenograph
}
\description{
Run Phenograph algorithm for clustering
}
\examples{
Use type markers to run Phenograph clustering for sce object with 1000 cells per sample,k set as 10
sce_rph<-runPheno(sce,k=10,downsample = 1000,seed = 1234, type_markers=T)
Use all markers and all cells for clustering, k set as 20
sce_rph<-runPheno(sce,k=20,downsample = NULL,seed=1234,type_markers=F)
}
