% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTest.R
\name{statTest}
\alias{statTest}
\title{Statistical test for cell frequencies}
\usage{
statTest(
  sce,
  abuntable = NULL,
  type = "DA",
  group,
  posthoc = T,
  method = "pairwise.wilcox.test"
)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{abuntable}{data frame generated from freqData function showing cell frequency data of each patient/sample,defalut is 'NULL'}

\item{type}{character string specifying the type of statistical analysis, with 'DA' stands for differential cell abundance analysis and 'DS'represents differential marker expression test facet by each cell subsets, default is 'DA'}

\item{group}{character string specifying the keyword for comparison, e.g.'condition'}

\item{posthoc}{logical vector; default is T which means post hoc analysis will be done after KW}

\item{method}{character string specifying the method used for post hoc analysis, default is 'pairwise.wilcox.test',another
option is 'dunn.test'}
}
\value{
Table(s) of the statistics with p value, method, and significance
}
\description{
Statistical comparison for cell frequencies/abundance in different study groups; Mann-Whitney test will be used
for two study groups, whereas Kruskal-Wallis test will be used for more than two groups with multiple testing correction and post hoc analysis.
}
\examples{
# Use sce object as input data, and specify method for post hoc analysis as dunn's test
# define the results as s object
s<-statTest(sce,group='outcome',posthoc=T,method='dunn.test')
# The function can automatically determine whether there are more than two groups, and apply KW and Pairwise Wilcoxon Rank Sum Tests by default
s<-statTest(sce,group='outcome')
# Input data can be a data frame instead of a sce object, just in case certain statistical modification has to be done before statistical comparison(e.g.remove certain outliers)
s<-statTest(abuntable=at,group='condition')
# Differential analysis of functional marker expression, choose type as DE, n=2 will perform Mann-Whitney test and n>=3 will perform pairwise wilcoxon test
test<-statTest(sce,type='DE',group='condition')

}
