% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotbcFreq.R
\name{plotbcFreq}
\alias{plotbcFreq}
\title{Plot total cell population frequencies after binary classification}
\usage{
plotbcFreq(sce, binary.results)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{binary_results}{list, results of binary classification}
}
\value{
a ggplot object
}
\description{
Visualisation of cell population frequencies/composition after binary classification, in the format of bar plot
}
\examples{
plotbcFreq(sce,binary.results)

}
