% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotbcHeatmap.R
\name{plotbcHeatmap}
\alias{plotbcHeatmap}
\title{Plot median marker expression heatmap of binary classification results}
\usage{
plotbcHeatmap(sce, binary.results, remove.unclass = FALSE)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{binary.results}{list, results from binaryClust function}

\item{remove.unclass}{logical vector, option to remove unclassified cells from plot, default is F}
}
\value{
a ggplot object
}
\description{
Visaulation of the median marker expression of each auto-annotated cluster
}
\examples{
plotbcHeatmap(sce,binary.results,remove.unclass=F)
}
